% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findWinningNeuron}
\alias{findWinningNeuron}
\title{Determine the Best Matching Unit (BMU) for a data vector}
\usage{
findWinningNeuron(weightMatrix, x, oldColumns)
}
\arguments{
\item{weightMatrix}{The weight matrix (codebook matrix).}

\item{x}{The vector.}

\item{oldColumns}{A boolean vector specifying the columns of the data vector which were already
available before the data set has been extended. In other words, the only
these columns are used for the calculation of the euclidian distance (for the BMU)
while the other other columns are ignored during this step.}
}
\value{
A list containing the following elements: The number of the row which is closest
to \code{x}, the x-coordinate and the y-coordinate of the found neuron in the quadratic
SOM grid.
}
\description{
This function takes a vector \code{x} and determines the row of the weight matrix which
minimizes the euclidian distance.
}
