% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{getDefaultNumberOfThreads}
\alias{getDefaultNumberOfThreads}
\title{Get Default Number of Threads}
\usage{
getDefaultNumberOfThreads()
}
\value{
The default number of threads.
}
\description{
The default value depends on the system.
}
