% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{setNumberOfThreads}
\alias{setNumberOfThreads}
\title{Set Number of Threads}
\usage{
setNumberOfThreads(n)
}
\arguments{
\item{n}{Number of threads.}
}
\description{
This function uses the function \code{setThreadOptions()} from the RcppParallel package.
}
\details{
See also the function \code{RcppParallel::setThreadOptions()} if you
want to change the stack size für the worker threads.
}
