% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsom.R
\name{som.init.extended}
\alias{som.init.extended}
\title{Codebook Matrix Initialization using PCA}
\usage{
som.init.extended(data, somSize, oldColumns)
}
\arguments{
\item{data}{The data set.}

\item{somSize}{The size of the resulting (quadratic) SOM grid of size \code{(somSize, somSize)}.}

\item{oldColumns}{A boolean vector specifying the columns of the data set which were already
available before the data set has been extended. In other words, the SOM initialization only
uses these columns for the calculation of Best Matching Units while the other other columns
are ignored during this step, but still modified during the update of the weight matrix}
}
\value{
The initialized codebook matrix.
}
\description{
This function provides an initial weight matrix for the SOM algorithm by applying
the principal component analysis (PCA) to a given data set, based on the
\code{som.init} function from the \code{som} package written by Jun Yan.
It also supports NA values and extended data sets.
}
