% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSom.R
\docType{package}
\name{xSom}
\alias{xSom}
\alias{xSom-package}
\title{xSom: A Parallel Version of the Som Algorithm for Extendable Datasets}
\description{
A parallelized version of the self-organizing map.
It also supports the extension of data sets where new columns
are appended without changing the resulting codebook matrix in the
old columns.
}
